Refactor "concurrent" tests to speed up build/link times
