+++ test/concurrent-downstream-done-one-eventual-fail.cc
@@ -1,0 +1,75 @@
+#include <deque>
+#include <string>
+
+#include "eventuals/callback.h"
+#include "eventuals/eventual.h"
+#include "eventuals/interrupt.h"
+#include "eventuals/iterate.h"
+#include "eventuals/let.h"
+#include "eventuals/map.h"
+#include "eventuals/reduce.h"
+#include "eventuals/terminal.h"
+#include "eventuals/then.h"
+#include "test/concurrent.h"
+
+using eventuals::Callback;
+using eventuals::Eventual;
+using eventuals::Interrupt;
+using eventuals::Iterate;
+using eventuals::Let;
+using eventuals::Map;
+using eventuals::Reduce;
+using eventuals::Terminate;
+using eventuals::Then;
+
+// Tests what happens when downstream is done before 'Concurrent()' is
+// done and one eventual fails.
+TYPED_TEST(ConcurrentTypedTest, DownstreamDoneOneEventualFail) {
+  std::deque<Callback<>> callbacks;
+
+  auto e = [&]() {
+    return Iterate({1, 2})
+        | this->ConcurrentOrConcurrentOrdered([&]() {
+            return Map(Let([&](int& i) {
+              return Eventual<std::string>()
+                  .interruptible()
+                  .start([&](auto& k, Interrupt::Handler& handler) mutable {
+                    if (i == 1) {
+                      callbacks.emplace_back([&k]() {
+                        k.Start("1");
+                      });
+                    } else {
+                      handler.Install([&k]() {
+                        k.Fail("error");
+                      });
+                      callbacks.emplace_back([]() {});
+                    }
+                  });
+            }));
+          })
+        | Reduce(
+               std::string(),
+               [](auto& result) {
+                 return Then([&](auto&& value) {
+                   result = value;
+                   return false; // Only take the first element!
+                 });
+               });
+  };
+
+  auto [future, k] = Terminate(e());
+
+  k.Start();
+
+  ASSERT_EQ(2, callbacks.size());
+
+  EXPECT_EQ(
+      std::future_status::timeout,
+      future.wait_for(std::chrono::seconds(0)));
+
+  for (auto& callback : callbacks) {
+    callback();
+  }
+
+  EXPECT_EQ("1", future.get());
+}
