+++ test/concurrent-emit-stop-interrupt.cc
@@ -1,0 +1,58 @@
+#include <string>
+#include <vector>
+
+#include "eventuals/collect.h"
+#include "eventuals/eventual.h"
+#include "eventuals/interrupt.h"
+#include "eventuals/let.h"
+#include "eventuals/map.h"
+#include "eventuals/stream.h"
+#include "eventuals/terminal.h"
+#include "test/concurrent.h"
+
+using eventuals::Collect;
+using eventuals::Eventual;
+using eventuals::Interrupt;
+using eventuals::Let;
+using eventuals::Map;
+using eventuals::Stream;
+using eventuals::Terminate;
+
+// Same as 'EmitFailInterrupt' except each eventual stops not fails.
+TYPED_TEST(ConcurrentTypedTest, EmitStopInterrupt) {
+  Interrupt interrupt;
+
+  auto e = [&]() {
+    return Stream<int>()
+               .interruptible()
+               .begin([](auto& k, Interrupt::Handler& handler) {
+                 handler.Install([&k]() {
+                   k.Stop();
+                 });
+                 k.Begin();
+               })
+               .next([i = 0](auto& k) mutable {
+                 i++;
+                 if (i == 1) {
+                   k.Emit(i);
+                 }
+               })
+        | this->ConcurrentOrConcurrentOrdered([&]() {
+            return Map(Let([&](int& i) {
+              return Eventual<std::string>([&](auto& k) {
+                k.Stop();
+                interrupt.Trigger();
+              });
+            }));
+          })
+        | Collect<std::vector<std::string>>();
+  };
+
+  auto [future, k] = Terminate(e());
+
+  k.Register(interrupt);
+
+  k.Start();
+
+  EXPECT_THROW(future.get(), eventuals::StoppedException);
+}
