+++ test/concurrent-interrupt-success.cc
@@ -1,0 +1,77 @@
+#include <deque>
+#include <string>
+#include <vector>
+
+#include "eventuals/callback.h"
+#include "eventuals/collect.h"
+#include "eventuals/eventual.h"
+#include "eventuals/interrupt.h"
+#include "eventuals/iterate.h"
+#include "eventuals/let.h"
+#include "eventuals/map.h"
+#include "eventuals/terminal.h"
+#include "test/concurrent.h"
+
+using eventuals::Callback;
+using eventuals::Collect;
+using eventuals::Eventual;
+using eventuals::Interrupt;
+using eventuals::Iterate;
+using eventuals::Let;
+using eventuals::Map;
+using eventuals::Terminate;
+
+// Tests that 'Concurrent()' and 'ConcurrentOrdered()' defers to the
+// eventuals on how to handle interrupts and in this case each eventual
+// ignores interrupts so we'll successfully collect all the values.
+TYPED_TEST(ConcurrentTypedTest, InterruptSuccess) {
+  std::deque<Callback<>> callbacks;
+
+  auto e = [&]() {
+    return Iterate({1, 2})
+        | this->ConcurrentOrConcurrentOrdered([&]() {
+            struct Data {
+              void* k;
+              int i;
+            };
+            return Map(Let([&](int& i) {
+              return Eventual<std::string>(
+                  [&, data = Data()](auto& k) mutable {
+                    using K = std::decay_t<decltype(k)>;
+                    data.k = &k;
+                    data.i = i;
+                    callbacks.emplace_back([&data]() {
+                      static_cast<K*>(data.k)->Start(std::to_string(data.i));
+                    });
+                  });
+            }));
+          })
+        | Collect<std::vector<std::string>>();
+  };
+
+  auto [future, k] = Terminate(e());
+
+  Interrupt interrupt;
+
+  k.Register(interrupt);
+
+  k.Start();
+
+  ASSERT_EQ(2, callbacks.size());
+
+  EXPECT_EQ(
+      std::future_status::timeout,
+      future.wait_for(std::chrono::seconds(0)));
+
+  interrupt.Trigger();
+
+  EXPECT_EQ(
+      std::future_status::timeout,
+      future.wait_for(std::chrono::seconds(0)));
+
+  for (auto& callback : callbacks) {
+    callback();
+  }
+
+  EXPECT_THAT(future.get(), this->OrderedOrUnorderedElementsAre("1", "2"));
+}
