+++ test/concurrent-moveable.cc
@@ -1,0 +1,41 @@
+#include <vector>
+
+#include "eventuals/collect.h"
+#include "eventuals/interrupt.h"
+#include "eventuals/iterate.h"
+#include "eventuals/let.h"
+#include "eventuals/map.h"
+#include "eventuals/terminal.h"
+#include "test/concurrent.h"
+
+using eventuals::Collect;
+using eventuals::Interrupt;
+using eventuals::Iterate;
+using eventuals::Let;
+using eventuals::Map;
+using eventuals::Terminate;
+
+// Tests that only moveable values will be moved into
+// 'Concurrent()' and 'ConcurrentOrdered()'.
+TYPED_TEST(ConcurrentTypedTest, Moveable) {
+  struct Moveable {
+    Moveable() = default;
+    Moveable(Moveable&&) = default;
+  };
+
+  auto e = [&]() {
+    return Iterate({Moveable()})
+        | this->ConcurrentOrConcurrentOrdered([]() {
+            return Map(Let([](auto& moveable) {
+              return 42;
+            }));
+          })
+        | Collect<std::vector<int>>();
+  };
+
+  auto [future, k] = Terminate(e());
+
+  k.Start();
+
+  EXPECT_THAT(future.get(), this->OrderedOrUnorderedElementsAre(42));
+}
