+++ test/concurrent-stream-fail.cc
@@ -1,0 +1,35 @@
+#include <string>
+#include <vector>
+
+#include "eventuals/collect.h"
+#include "eventuals/map.h"
+#include "eventuals/stream.h"
+#include "eventuals/terminal.h"
+#include "test/concurrent.h"
+
+using eventuals::Collect;
+using eventuals::Map;
+using eventuals::Stream;
+using eventuals::Terminate;
+
+// Tests when when upstream fails the result will be fail.
+TYPED_TEST(ConcurrentTypedTest, StreamFail) {
+  auto e = [&]() {
+    return Stream<int>()
+               .next([](auto& k) {
+                 k.Fail("error");
+               })
+        | this->ConcurrentOrConcurrentOrdered([]() {
+            return Map([](int i) {
+              return std::to_string(i);
+            });
+          })
+        | Collect<std::vector<std::string>>();
+  };
+
+  auto [future, k] = Terminate(e());
+
+  k.Start();
+
+  EXPECT_THROW(future.get(), std::exception_ptr);
+}
